Move particles into their own package
