+++ radixdlt-java/src/main/java/com/radixdlt/client/application/translate/AddressTokenReducer.java
@@ -3,10 +3,10 @@ package com.radixdlt.client.application.translate;
 import com.radixdlt.client.application.objects.Amount;
 import com.radixdlt.client.application.objects.Token;
 import com.radixdlt.client.core.address.RadixAddress;
-import com.radixdlt.client.core.atoms.AtomFeeConsumable;
-import com.radixdlt.client.core.atoms.Consumable;
+import com.radixdlt.client.core.atoms.particles.AtomFeeConsumable;
+import com.radixdlt.client.core.atoms.particles.Consumable;
 import com.radixdlt.client.core.atoms.RadixHash;
-import com.radixdlt.client.core.atoms.Spin;
+import com.radixdlt.client.core.atoms.particles.Spin;
 import com.radixdlt.client.core.ledger.ParticleStore;
 import io.reactivex.Observable;
 import java.util.HashMap;
