+++ radixdlt-java/src/main/java/com/radixdlt/client/application/translate/AddressTokenState.java
@@ -1,7 +1,7 @@
 package com.radixdlt.client.application.translate;
 
 import com.radixdlt.client.application.objects.Amount;
-import com.radixdlt.client.core.atoms.Consumable;
+import com.radixdlt.client.core.atoms.particles.Consumable;
 import java.util.List;
 
 public class AddressTokenState {
