+++ radixdlt-java/src/main/java/com/radixdlt/client/application/translate/DataStoreTranslator.java
@@ -6,8 +6,8 @@ import com.radixdlt.client.application.actions.DataStore;
 import com.radixdlt.client.application.objects.Data;
 import com.radixdlt.client.core.atoms.Atom;
 import com.radixdlt.client.core.atoms.AtomBuilder;
-import com.radixdlt.client.core.atoms.DataParticle;
-import com.radixdlt.client.core.atoms.DataParticle.DataParticleBuilder;
+import com.radixdlt.client.core.atoms.particles.DataParticle;
+import com.radixdlt.client.core.atoms.particles.DataParticle.DataParticleBuilder;
 import com.radixdlt.client.core.atoms.Payload;
 import com.radixdlt.client.core.crypto.EncryptedPrivateKey;
 import com.radixdlt.client.core.crypto.Encryptor;
