+++ radixdlt-java/src/main/java/com/radixdlt/client/core/atoms/AtomBuilder.java
@@ -1,5 +1,8 @@
 package com.radixdlt.client.core.atoms;
 
+import com.radixdlt.client.core.atoms.particles.AtomFeeConsumable;
+import com.radixdlt.client.core.atoms.particles.ChronoParticle;
+import com.radixdlt.client.core.atoms.particles.Particle;
 import com.radixdlt.client.core.crypto.ECPublicKey;
 import java.util.ArrayList;
 import java.util.List;
