+++ radixdlt-java/src/main/java/com/radixdlt/client/core/atoms/particles/AtomFeeConsumable.java
@@ -1,6 +1,7 @@
-package com.radixdlt.client.core.atoms;
+package com.radixdlt.client.core.atoms.particles;
 
 import com.radixdlt.client.core.address.EUID;
+import com.radixdlt.client.core.atoms.AccountReference;
 import java.util.List;
 
 public class AtomFeeConsumable extends Consumable {
