+++ radixdlt-java/src/main/java/com/radixdlt/client/core/atoms/particles/Consumable.java
@@ -1,8 +1,10 @@
-package com.radixdlt.client.core.atoms;
+package com.radixdlt.client.core.atoms.particles;
 
 import com.google.gson.annotations.SerializedName;
 import com.radixdlt.client.core.TokenClassReference;
 import com.radixdlt.client.core.address.EUID;
+import com.radixdlt.client.core.atoms.AccountReference;
+import com.radixdlt.client.core.atoms.RadixHash;
 import com.radixdlt.client.core.crypto.ECKeyPair;
 import com.radixdlt.client.core.crypto.ECPublicKey;
 import com.radixdlt.client.core.serialization.Dson;
