+++ radixdlt-java/src/main/java/com/radixdlt/client/core/atoms/particles/UniqueParticle.java
@@ -1,6 +1,7 @@
-package com.radixdlt.client.core.atoms;
+package com.radixdlt.client.core.atoms.particles;
 
 import com.radixdlt.client.core.address.EUID;
+import com.radixdlt.client.core.atoms.Payload;
 import com.radixdlt.client.core.crypto.ECKeyPair;
 import com.radixdlt.client.core.crypto.ECPublicKey;
 import java.util.Collections;
