+++ radixdlt-java/src/main/java/com/radixdlt/client/core/ledger/ConsumableDataSource.java
@@ -2,7 +2,7 @@ package com.radixdlt.client.core.ledger;
 
 import com.radixdlt.client.core.address.RadixAddress;
 import com.radixdlt.client.core.atoms.Atom;
-import com.radixdlt.client.core.atoms.Consumable;
+import com.radixdlt.client.core.atoms.particles.Consumable;
 import io.reactivex.Observable;
 import java.util.concurrent.ConcurrentHashMap;
