+++ radixdlt-java/src/main/java/com/radixdlt/client/core/ledger/ParticleStore.java
@@ -1,7 +1,7 @@
 package com.radixdlt.client.core.ledger;
 
 import com.radixdlt.client.core.address.RadixAddress;
-import com.radixdlt.client.core.atoms.Consumable;
+import com.radixdlt.client.core.atoms.particles.Consumable;
 import io.reactivex.Observable;
 
 public interface ParticleStore {
