+++ radixdlt-java/src/main/java/com/radixdlt/client/core/ledger/RadixAtomValidator.java
@@ -5,7 +5,7 @@ import com.radixdlt.client.core.atoms.Atom;
 import com.radixdlt.client.core.atoms.AtomValidationException;
 import com.radixdlt.client.core.atoms.AtomValidator;
 import com.radixdlt.client.core.atoms.RadixHash;
-import com.radixdlt.client.core.atoms.Spin;
+import com.radixdlt.client.core.atoms.particles.Spin;
 import com.radixdlt.client.core.crypto.ECSignature;
 import java.util.Objects;
 import java.util.Optional;
