+++ radixdlt-java/src/main/java/com/radixdlt/client/core/serialization/RadixJson.java
@@ -2,8 +2,8 @@ package com.radixdlt.client.core.serialization;
 
 import com.radixdlt.client.core.TokenClassReference;
 import com.radixdlt.client.core.atoms.AccountReference;
-import com.radixdlt.client.core.atoms.TokenParticle;
-import com.radixdlt.client.core.atoms.Spin;
+import com.radixdlt.client.core.atoms.particles.TokenParticle;
+import com.radixdlt.client.core.atoms.particles.Spin;
 import java.io.IOException;
 import java.lang.reflect.Type;
 import java.util.HashMap;
@@ -32,13 +32,13 @@ import com.google.gson.stream.JsonWriter;
 import com.radixdlt.client.core.address.EUID;
 import com.radixdlt.client.core.address.RadixUniverseType;
 import com.radixdlt.client.core.atoms.Atom;
-import com.radixdlt.client.core.atoms.AtomFeeConsumable;
-import com.radixdlt.client.core.atoms.ChronoParticle;
-import com.radixdlt.client.core.atoms.Consumable;
-import com.radixdlt.client.core.atoms.DataParticle;
-import com.radixdlt.client.core.atoms.Emission;
+import com.radixdlt.client.core.atoms.particles.AtomFeeConsumable;
+import com.radixdlt.client.core.atoms.particles.ChronoParticle;
+import com.radixdlt.client.core.atoms.particles.Consumable;
+import com.radixdlt.client.core.atoms.particles.DataParticle;
+import com.radixdlt.client.core.atoms.particles.Emission;
 import com.radixdlt.client.core.atoms.MetadataMap;
-import com.radixdlt.client.core.atoms.Particle;
+import com.radixdlt.client.core.atoms.particles.Particle;
 import com.radixdlt.client.core.atoms.Payload;
 import com.radixdlt.client.core.crypto.ECKeyPair;
 import com.radixdlt.client.core.crypto.ECPublicKey;
