+++ radixdlt-java/src/test/java/com/radixdlt/client/application/translate/AddressTokenReducerTest.java
@@ -6,9 +6,9 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import com.radixdlt.client.core.address.RadixAddress;
-import com.radixdlt.client.core.atoms.Consumable;
+import com.radixdlt.client.core.atoms.particles.Consumable;
 import com.radixdlt.client.core.atoms.RadixHash;
-import com.radixdlt.client.core.atoms.Spin;
+import com.radixdlt.client.core.atoms.particles.Spin;
 import com.radixdlt.client.core.ledger.ParticleStore;
 import io.reactivex.Observable;
 import io.reactivex.observers.TestObserver;
