+++ radixdlt-java/src/test/java/com/radixdlt/client/application/translate/TransactionAtomsTest.java
@@ -8,8 +8,8 @@ import com.radixdlt.client.application.objects.Token;
 import com.radixdlt.client.core.address.RadixAddress;
 import com.radixdlt.client.core.atoms.AccountReference;
 import com.radixdlt.client.core.atoms.Atom;
-import com.radixdlt.client.core.atoms.Consumable;
-import com.radixdlt.client.core.atoms.Spin;
+import com.radixdlt.client.core.atoms.particles.Consumable;
+import com.radixdlt.client.core.atoms.particles.Spin;
 import com.radixdlt.client.core.crypto.ECKeyPair;
 import com.radixdlt.client.core.crypto.ECPublicKey;
 import io.reactivex.observers.TestObserver;
